Move User class to own file
